(re)introduce lz4 on all platforms (so rocksdb can use it)
On arch linux systemd pulls in lz4 (including headers et al) and rocksdb finds and uses it.
If lz4-dev(el) package is installed on the other platforms, rocksdb will also use it there.
The extra dependency should (automatically) be picked up by the package building in the volumedriver repo.