release: introduce BOOST_PYTHON_LIB_NAME + build config for 16.04 / Py3.5
On Ubuntu 16.04 boost python comes in 2 flavours - -lboost_python-py27
(default -lboost_python symlinks to it) and -lboost_python-py35.
The latter is to be used, but since these names are 'debian-isms', a new
configurable is introduced for the release build which allows to set the
boost python library name in the build configuration.